refactor: move WasmReader
