+++ src/core/decoding/mod.rs
@@ -1,3 +1,5 @@
 pub mod error;
 pub mod modules;
+pub mod reader;
 pub mod types;
+pub mod values;
