+++ src/core/decoding/modules/mod.rs
@@ -1,4 +1,4 @@
-use crate::core::{decoding::error::DecodingError, reader::WasmReader};
+use crate::core::decoding::{error::DecodingError, reader::WasmReader};
 
 pub mod section_header;
