+++ src/core/mod.rs
@@ -1,7 +1,6 @@
 pub mod decoding;
 pub mod error;
 pub mod fixed_capacity_vec;
-pub mod reader;
 pub mod rw_spinlock;
 pub mod sidetable;
 pub mod structure;
