+++ src/core/structure/modules/data_segments.rs
@@ -2,7 +2,7 @@ use core::fmt;
 
 use alloc::{format, vec::Vec};
 
-use crate::core::{reader::span::Span, structure::modules::indices::MemIdx};
+use crate::core::{decoding::reader::span::Span, structure::modules::indices::MemIdx};
 
 #[derive(Clone)]
 pub struct DataSegment {
