+++ src/validation/read_constant_expression.rs
@@ -1,7 +1,7 @@
 use alloc::vec::Vec;
 
-use crate::core::reader::span::Span;
-use crate::core::reader::WasmReader;
+use crate::core::decoding::reader::span::Span;
+use crate::core::decoding::reader::WasmReader;
 use crate::core::structure::modules::indices::{FuncIdx, IdxVec, TypeIdx};
 use crate::core::structure::types::GlobalType;
 use crate::core::utils::ToUsizeExt;
